# PERF: DataFrame.values for pyarrow-backed numeric types