@lukemanley 2023-04-02
> This also changes behavior (e.g. getting float instead of object). 

Yes, the performance improvement is due to avoiding the cast to object. Note, this behavior actually already exists on `main` for a `DataFrame` with a single column:

Behavior on `main`:
```
import pandas as pd

df1 = pd.DataFrame({"a": [1.0, 2.0, 3.0]}, dtype="float64[pyarrow]")
df2 = pd.DataFrame({"a": [1.0, 2.0, pd.NA]}, dtype="float64[pyarrow]")

print(df1.values.dtype)  # float64
print(df2.values.dtype)  # object
```

> Personally, I think this is fine but we have an unresolved discussion about this somewhere. We should decide there first before special casing here I'd say

Sure, I think you might be referring to https://github.com/pandas-dev/pandas/issues/22791