@lukemanley 2023-04-09
closing for now pending further discussion in https://github.com/pandas-dev/pandas/issues/22791